extract stories into individual modules
