+++ packages/checkbox/stories/default.tsx
@@ -1,0 +1,24 @@
+import React from "react"
+
+import { CheckboxGroup, Checkbox } from "../index"
+
+/* eslint-disable react/jsx-key */
+const checkboxes = [
+	<Checkbox label="Guardian Today: UK" value="today_uk" />,
+	<Checkbox label="Guardian Today: US" value="today_us" />,
+]
+/* eslint-enable react/jsx-key */
+
+const defaultLight = () => (
+	<CheckboxGroup name="emails">
+		{checkboxes.map((checkbox, index) =>
+			React.cloneElement(checkbox, { key: index }),
+		)}
+	</CheckboxGroup>
+)
+
+defaultLight.story = {
+	name: "default light",
+}
+
+export { defaultLight }
