+++ packages/checkbox/stories/indeterminate.tsx
@@ -1,0 +1,17 @@
+import React from "react"
+
+import { Checkbox } from "../index"
+
+const indeterminateLight = () => (
+	<Checkbox
+		indeterminate={true}
+		value="indeterminate"
+		label="Indeterminate"
+	/>
+)
+
+indeterminateLight.story = {
+	name: "indeterminate light",
+}
+
+export { indeterminateLight }
